+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## Version 1.4.10 - Community 1.0.0
+    * Loans & Payments
+        * [CP-4057] - Improve Rahisi Model UI to show more details and make it more user friendly
+
 ## Version 1.4.9 - Community 1.0.0
     * Bulk Uploads
         * [CP-4042] - Add bulk upload to remove clients from groups